# AppWebCanvan/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Kanban API",
        "description": "API para gestionar tableros Kanban con columnas y tareas",
        "contact": {
            "email": "admin@kanban.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obtener todas las tareas",
                "description": "Retorna una lista de todas las tareas",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Lista de tareas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Implementar login"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Crear sistema de autenticación"
                                            },
                                            "column_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "fecha_asignacion": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-10-05"
                                            },
                                            "fecha_limite": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-10-15"
                                            },
                                            "asignador": {
                                                "type": "string",
                                                "example": "Tech Lead"
                                            },
                                            "responsable": {
                                                "type": "string",
                                                "example": "Juan Pérez"
                                            },
                                            "avance": {
                                                "type": "integer",
                                                "maximum": 100,
                                                "minimum": 0,
                                                "example": 75
                                            },
                                            "prioridad": {
                                                "type": "string",
                                                "enum": [
                                                    "baja",
                                                    "media",
                                                    "alta",
                                                    "urgente"
                                                ],
                                                "example": "alta"
                                            },
                                            "total_days": {
                                                "description": "Días totales para completar la tarea",
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "missing_days": {
                                                "description": "Días restantes para la fecha límite",
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "status": {
                                                "description": "Estado calculado basado en fechas y avance",
                                                "type": "string",
                                                "example": "En progreso"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear una nueva tarea",
                "description": "Crea una nueva tarea en una columna específica",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "column_id",
                                    "fecha_asignacion",
                                    "fecha_limite",
                                    "asignador",
                                    "responsable",
                                    "prioridad"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Nueva tarea"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción detallada de la tarea"
                                    },
                                    "column_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_asignacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-05"
                                    },
                                    "fecha_limite": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    },
                                    "asignador": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Tech Lead"
                                    },
                                    "responsable": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Juan Pérez"
                                    },
                                    "avance": {
                                        "type": "integer",
                                        "maximum": 100,
                                        "minimum": 0,
                                        "example": 0
                                    },
                                    "prioridad": {
                                        "type": "string",
                                        "enum": [
                                            "baja",
                                            "media",
                                            "alta",
                                            "urgente"
                                        ],
                                        "example": "media"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Nueva tarea"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción de la tarea"
                                        },
                                        "column_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "fecha_asignacion": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-10-05"
                                        },
                                        "fecha_limite": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-10-15"
                                        },
                                        "asignador": {
                                            "type": "string",
                                            "example": "Tech Lead"
                                        },
                                        "responsable": {
                                            "type": "string",
                                            "example": "Juan Pérez"
                                        },
                                        "avance": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "prioridad": {
                                            "type": "string",
                                            "example": "media"
                                        },
                                        "total_days": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "missing_days": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Pendiente"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo nombre es obligatorio."
                                                    }
                                                },
                                                "prioridad": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "La prioridad seleccionada es inválida."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}